+++ BBTAG_IM.vcxproj
@@ -159,7 +159,7 @@
     <ClCompile Include="src\Web\url_downloader.cpp" />
     <ClCompile Include="src\WindowManager\CustomHud.cpp" />
     <ClCompile Include="src\WindowManager\fonts.cpp" />
-    <ClCompile Include="src\WindowManager\HitboxOverlay.cpp" />
+    <ClCompile Include="src\WindowManager\WorldToScreen.cpp" />
     <ClCompile Include="src\WindowManager\PaletteEditor.cpp" />
     <ClCompile Include="src\WindowManager\WindowManager.cpp" />
   </ItemGroup>
@@ -211,7 +211,7 @@
     <ClInclude Include="src\Web\url_downloader.h" />
     <ClInclude Include="src\WindowManager\CustomHud.h" />
     <ClInclude Include="src\WindowManager\fonts.h" />
-    <ClInclude Include="src\WindowManager\HitboxOverlay.h" />
+    <ClInclude Include="src\WindowManager\WorldToScreen.h" />
     <ClInclude Include="src\WindowManager\PaletteEditor.h" />
     <ClInclude Include="src\WindowManager\WindowManager.h" />
   </ItemGroup>
